LayoutTests/js/math-denorm-expected.txt

This test checks the behavior of the Math.ceil and Math.floor on denormalized FP values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Math.ceil(Number.MIN_VALUE) is 1
PASS Math.ceil(-Number.MIN_VALUE) is -0
PASS Math.floor(Number.MIN_VALUE) is 0
PASS Math.floor(-Number.MIN_VALUE) is -1
PASS successfullyParsed is true

TEST COMPLETE
